@@ -1,17 +1,14 @@
 package ru.skypro.homework.entity;
 
-import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.Getter;
-import lombok.Setter;
+import jakarta.persistence.Table;
 import ru.skypro.homework.dto.Role;
 
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.validation.constraints.NotNull;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.validation.constraints.NotNull;
 
 @Entity
-@Getter
-@Setter
+@Table(name = "users")
 public class User {
     @Id
     @NotNull
@@ -26,10 +23,11 @@ public class User {
     private String phone;
     @NotNull
     private Role role;
-    @NotNull
     private String image;
+    @NotNull
+    private String password;
 
-    public User(int id, String email, String firstName, String lastName, String phone, Role role, String image) {
+    public User(int id, String email, String firstName, String lastName, String phone, Role role, String image, String password) {
         this.id = id;
         this.email = email;
         this.firstName = firstName;
@@ -37,7 +35,73 @@ public class User {
         this.phone = phone;
         this.role = role;
         this.image = image;
+        this.password = password;
     }
 
     public User() {}
+
+    @NotNull
+    public int getId() {
+        return id;
+    }
+
+    public void setId(@NotNull int id) {
+        this.id = id;
+    }
+
+    public @NotNull String getEmail() {
+        return email;
+    }
+
+    public void setEmail(@NotNull String email) {
+        this.email = email;
+    }
+
+    public @NotNull String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(@NotNull String firstName) {
+        this.firstName = firstName;
+    }
+
+    public @NotNull String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(@NotNull String lastName) {
+        this.lastName = lastName;
+    }
+
+    public @NotNull String getPhone() {
+        return phone;
+    }
+
+    public void setPhone(@NotNull String phone) {
+        this.phone = phone;
+    }
+
+    public @NotNull Role getRole() {
+        return role;
+    }
+
+    public void setRole(@NotNull Role role) {
+        this.role = role;
+    }
+
+    public @NotNull String getImage() {
+        return image;
+    }
+
+    public void setImage(@NotNull String image) {
+        this.image = image;
+    }
+
+    public @NotNull String getPassword() {
+        return password;
+    }
+
+    public void setPassword(@NotNull String password) {
+        this.password = password;
+    }
 }
