@@ -4,7 +4,6 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 import ru.skypro.homework.dto.NewPassword;
 import ru.skypro.homework.dto.UpdateUser;
@@ -13,10 +12,13 @@ import ru.skypro.homework.service.UserService;
 
 @RestController
 @RequestMapping("/user")
-@RequiredArgsConstructor
 @Schema(description = "Контроллер для работы с пользователями")
 public class UserController {
     private final UserService userService;
+
+    public UserController(UserService userService) {
+        this.userService = userService;
+    }
 
     @PatchMapping("/set_password")
     @Operation(summary = "Set new password", description = "Обновление пароля",
