@@ -3,5 +3,5 @@ package ru.skypro.homework.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.skypro.homework.entity.Ad;
 
-public interface AdRepository extends JpaRepository<Integer, Ad> {
+public interface AdRepository extends JpaRepository<Ad, Integer> {
 }
