@@ -3,5 +3,5 @@ package ru.skypro.homework.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.skypro.homework.entity.Comment;
 
-public interface CommentRepository extends JpaRepository<Integer, Comment> {
+public interface CommentRepository extends JpaRepository<Comment, Integer> {
 }
